fix: increase barito_app cost related column int limit

+++ db/migrate/20230223100302_add_latest_cost_to_barito_app.rb
@@ -1,6 +1,6 @@
 class AddLatestCostToBaritoApp < ActiveRecord::Migration[5.2]
   def change
-    add_column :barito_apps, :latest_cost, :integer
-    add_column :barito_apps, :latest_ingested_log_bytes, :integer
+    add_column :barito_apps, :latest_cost, :bigint
+    add_column :barito_apps, :latest_ingested_log_bytes, :bigint
   end
 end
